fix(salt): few fixes; freebsd support

+++ salt/salter.sh
@@ -28,12 +28,13 @@ RC=0
 BASE=/srv
 BASE_ETC=/etc
 STATEDIR=''
-if [ `uname` == 'FreeBSD' ]; then
+if [ `uname` == "FreeBSD" ]; then
     BASE=/usr/local/etc
     BASE_ETC=/usr/local/etc
     STATEDIR=states
+    SUBDIR=salt
 fi
-PILLARFS=${BASE:-/srv}/pillar
+PILLARFS=${BASE:-/srv}/${SUBDIR}/pillar
 SALTFS=${BASE:-/srv}/salt/${STATEDIR}
 SKIP_UNNECESSARY_CLONE=''
 HOMEBREW=/usr/local/bin/brew
@@ -70,6 +71,8 @@ pkg-query() {
          /usr/bin/dpkg-query --list | grep ${PACKAGE}
     elif [ -f "/usr/bin/pacman" ]; then
          /usr/bin/pacman -Qi ${PACKAGE}
+    elif [[ -f "/usr/sbin/pkg" ]]; then
+         /usr/sbin/pkg query ${PACKAGE}
     fi
 }
 
@@ -91,7 +94,8 @@ pkg-install() {
              fi
              ;;
 
-    linux*)  if [ -f "/usr/bin/zypper" ]; then
+    linux*|freebsd*)
+             if [ -f "/usr/bin/zypper" ]; then
                  /usr/bin/zypper update -y || exit 1
                  /usr/bin/zypper --non-interactive install ${PACKAGES} || exit 1
              elif [ -f "/usr/bin/emerge" ]; then
@@ -103,14 +107,17 @@ pkg-install() {
              elif [ -f "/usr/bin/dnf" ]; then
                  /usr/bin/dnf install -y --best --allowerasing ${PACKAGES} || exit 1
              elif [ -f "/usr/bin/yum" ]; then
-                 /usr/bin/yum update -y || exit 1 
-                 /usr/bin/yum install -y ${PACKAGES} || exit 1 
+                 /usr/bin/yum update -y || exit 1
+                 /usr/bin/yum install -y ${PACKAGES} || exit 1
              elif [[ -f "/usr/bin/apt-get" ]]; then
                  /usr/bin/apt-get update --fix-missing -y || exit 1
                  /usr/bin/apt-add-repository universe
                  /usr/bin/apt autoremove -y
                  /usr/bin/apt-get update -y
                  /usr/bin/apt-get install -y ${PACKAGES} || exit 1
+             elif [[ -f "/usr/sbin/pkg" ]]; then
+                 /usr/sbin/pkg update -f --quiet || exit 1
+                 /usr/sbin/pkg install --automatic --yes ${PACKAGES} || exit 1
              fi
     esac
 }
@@ -131,9 +138,11 @@ pkg-update() {
              elif [ -f "/usr/bin/dnf" ]; then
                  /usr/bin/dnf upgrade -y --allowerasing ${PACKAGES} || exit 1
              elif [ -f "/usr/bin/yum" ]; then
-                 /usr/bin/yum update -y ${PACKAGES} || exit 1 
+                 /usr/bin/yum update -y ${PACKAGES} || exit 1
              elif [[ -f "/usr/bin/apt-get" ]]; then
                  /usr/bin/apt-get upgrade -y ${PACKAGES} || exit 1
+             elif [[ -f "/usr/sbin/pkg" ]]; then
+                 /usr/sbin/pkg upgrade --yes ${PACKAGES} || exit 1
              fi
     esac
 }
@@ -145,7 +154,8 @@ pkg-remove() {
                  su - ${USER} -c "${HOMEBREW} uninstall ${p} --force"
              done
              ;;
-    linux*)  if [ -f "/usr/bin/zypper" ]; then
+    linux*|freebsd*)
+             if [ -f "/usr/bin/zypper" ]; then
                  /usr/bin/zypper --non-interactive rm ${PACKAGES} || exit 1
              elif [ -f "/usr/bin/emerge" ]; then
                  /usr/bin/emerge -C ${PACKAGES} || exit 1
@@ -154,9 +164,11 @@ pkg-remove() {
              elif [ -f "/usr/bin/dnf" ]; then
                  /usr/bin/dnf remove -y ${PACKAGES} || exit 1
              elif [ -f "/usr/bin/yum" ]; then
-                 /usr/bin/yum remove -y ${PACKAGES} || exit 1 
+                 /usr/bin/yum remove -y ${PACKAGES} || exit 1
              elif [[ -f "/usr/bin/apt-get" ]]; then
                  /usr/bin/apt-get remove -y ${PACKAGES} || exit 1
+             elif [[ -f "/usr/sbin/pkg" ]]; then
+                 /usr/sbin/pkg delete --yes ${PACKAGES} || exit 1
              fi
     esac
 }
@@ -228,8 +240,9 @@ salt-bootstrap() {
              export HOMEBREW_CURLRC=1
              ;;
 
-     linux*) pkg-update 2>/dev/null
-             echo "Setup Linux baseline and install saltstack masterless minion ..."
+     linux*|freebsd*)
+             pkg-update 2>/dev/null
+             echo "Setup Linux/FreeBSD baseline and install saltstack masterless minion ..."
              if [ -f "/usr/bin/dnf" ]; then
                  PACKAGES="--best --allowerasing git wget redhat-rpm-config"
              elif [ -f "/usr/bin/yum" ]; then
@@ -240,6 +253,8 @@ salt-bootstrap() {
                  PACKAGES="git ssh wget curl software-properties-common"
              elif [ -f "/usr/bin/pacman" ]; then
                  PACKAGES="git wget psutils"
+             elif [ -f "/usr/sbin/pkg" ]; then
+                 PACKAGES="git wget psutils"
              fi
              pkg-install ${PACKAGES} 2>/dev/null
              if (( $? > 0 )); then
@@ -249,6 +264,7 @@ salt-bootstrap() {
              wget -O install_salt.sh https://bootstrap.saltstack.com || exit 10
              (sh install_salt.sh -x python3 ${SALT_VERSION} && rm -f install_salt.sh) || exit 10
              rm -f install_salt.sh 2>/dev/null
+             ;;
     esac
     ### stop debian interference with services (https://wiki.debian.org/chroot)
     if [ -f "/usr/bin/apt-get" ]; then
@@ -260,22 +276,30 @@ EOF
         ### Enforce python3
         rm /usr/bin/python 2>/dev/null; ln -s /usr/bin/python3 /usr/bin/python
     fi
-    ### install salt-api (except arch/macos)
-    [ ! -f "/etc/arch-release" ] && [ "$(uname)" != "Darwin" ] && pkg-install salt-api
+    ### install salt-api (except arch/macos/freebsd)
+    [ ! -f "/etc/arch-release" ] && [ "$(uname)" != "Darwin" ] && [ "$(uname)" != "FreeBSD" ] && pkg-install salt-api
 
-    ### salt services
+    ### salt minion
+    [ ! -f "${BASE_ETC}/salt/minion" ] && echo "File ${BASE_ETC}/salt/minion not found" && exit 1
     if [[ "`uname`" == "FreeBSD" ]] || [[ "`uname`" == "Darwin" ]]; then
         sed -i"bak" "s@^\s*#*\s*master\s*: salt\s*\$@master: ${solution[saltmaster]}@" ${BASE_ETC}/salt/minion
     else
         sed -i "s@^\s*#*\s*master\s*: salt\s*\$@master: ${solution[saltmaster]}@" ${BASE_ETC}/salt/minion
     fi
+    ### salt services
     (systemctl enable salt-api && systemctl start salt-api) 2>/dev/null || service start salt-api 2>/dev/null
     (systemctl enable salt-master && systemctl start salt-master) 2>/dev/null || service start salt-master 2>/dev/null
     (systemctl enable salt-minion && systemctl start salt-minion) 2>/dev/null || service start salt-minion 2>/dev/null
     salt-key -A --yes >/dev/null 2>&1     ##accept pending registrations
+
+    ### reboot to activate a new kernel?
     echo && KERNEL_VERSION=$( uname -r | awk -F. '{print $1"."$2"."$3"."$4"."$5}' )
     echo "kernel before: ${KERNEL_VERSION}"
-    echo "kernel after: $( pkg-query linux 2>/dev/null )"
+    if [ "$(uname)" == "FreeBSD" ]; then
+        echo "kernel after: $( /bin/freebsd-version -k 2>/dev/null )"
+    else
+        echo "kernel after: $( pkg-query linux 2>/dev/null )"
+    fi
     echo "Reboot if kernel was major-upgraded; if unsure reboot!"
     echo
 }
@@ -306,6 +330,8 @@ gitclone() {
     else
         git clone ${URI}/${ENTITY}/${REPO} ${SALTFS}/namespaces/${ENTITY}/${REPO} >/dev/null 2>&1 || exit 11
     fi
+    ## Its important to ensure symlink points to *this* formula
+    rm -f ${SALTFS}/${ALIAS} 2>/dev/null ## this is important make sure symlink is current
     echo && ln -s ${SALTFS}/namespaces/${ENTITY}/${REPO}/${SUBDIR} ${SALTFS}/${ALIAS} 2>/dev/null
 }
 
@@ -315,7 +341,7 @@ highstate() {
     ## prepare states
     ACTION=${1} && STATEDIR=${2} && TARGET=${3}
     for PROFILE in ${solution[saltdir]}/${ACTION}/${TARGET} ${your[saltdir]}/${ACTION}/${TARGET}
-    do  
+    do
         [ -f ${PROFILE}.sls ] && cp ${PROFILE}.sls ${SALTFS}/top.sls && break
         [ -f ${PROFILE}/init.sls ] && cp ${PROFILE}/init.sls ${SALTFS}/top.sls && break
     done
@@ -366,20 +392,20 @@ usage() {
     echo "  [-u <username>]" 1>&2
     echo "        A Loginname (current or corporate or root user)." 1>&2
     echo "        its mandatory for some Linux profiles" 1>&2
-    echo "        but not required on MacOS" 1>&2 
+    echo "        but not required on MacOS" 1>&2
     echo 1>&2
     echo "  TARGETS" 1>&2
     echo 1>&2
-    echo "\tbootstrap\t\tRun salt-bootstrap with additions" 1>&2
+    echo -e "\tbootstrap\t\tRun salt-bootstrap with additions" 1>&2
     echo 1>&2
-    echo "\tsalter\t\tInstall salter and salt-formula" 1>&2
+    echo -e "\tsalter\t\tInstall salter and salt-formula" 1>&2
     echo 1>&2
     if [ "${solution[entity]}" != "salter" ]; then
-       echo "\t${solution[entity]}\tDeploy ${solution[repo]}" 1>&2
+       echo -e "\t${solution[entity]}\tDeploy ${solution[repo]}" 1>&2
        echo 1>&2
     fi
     echo " ${solution[targets]}" 1>&2
-    echo "\t\t\tApply specific ${solution[repo]} state" 1>&2
+    echo -e "\t\t\tApply specific ${solution[repo]} state" 1>&2
     echo 1>&2
     echo "  OPTIONS" 1>&2
     echo 1>&2
@@ -458,7 +484,6 @@ developer-definitions() {
 }
 
 solution-definitions() {
-    ### solution polyrepo or monorepo ###
     solution['saltmaster']=""
     solution['uri']="https://github.com"
     solution['entity']="opensds"
@@ -480,7 +505,8 @@ solution-definitions() {
 }
 
 custom-install() {
-    ### required - salter-engine is insufficient ###
+    echo
+    ### required if salter-engine is insufficient ###
     gitclone ${solution[uri]} ${solution[entity]} ${solution[repo]} ${solution[alias]} ${solution[subdir]}
     SKIP_UNNECESSARY_CLONE='true'
     losetup -D 2>/dev/null
